# docs/swagger.yaml
basePath: /
host: localhost:4123
definitions:
...
  model.DockerContainer:
    properties:
      command:
        type: string
      created_at:
        type: string
      id:
        type: string
      image:
        type: string
      labels:
        type: string
      names:
        items:
          type: string
        type: array
      ports:
        type: string
      status:
        type: string
    type: object
...
  model.Server:
    properties:
      auth_type:
        type: string
      created_at:
        type: string
      description:
        type: string
      host:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      private_key:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  types.PageResponse:
    properties:
      records: {}
      total:
        type: integer
    type: object
...
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Minimal OpenAPI spec for Zebra-CICD endpoints
  title: Zebra-CICD API
  version: 0.1.0
paths:
...
  /api/linux-machines:
    get:
      description: 获取所有服务器的列表，支持按名称、主机等条件查询
      parameters:
      - description: 服务器名称
        in: query
        name: name
        type: string
      - description: 服务器主机地址
        in: query
        name: host
        type: string
      - description: 是否激活
        in: query
        name: isActive
        type: boolean
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 页数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/types.PageResponse'
                  - properties:
                      records:
                        items:
                          $ref: '#/definitions/model.Server'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取服务器列表
      tags:
      - linux-machines
    post:
      consumes:
      - application/json
      description: 创建一个新的服务器连接信息
      parameters:
      - description: 服务器信息
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/model.Server'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Server'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建服务器
      tags:
      - linux-machines
  /api/linux-machines/{id}:
    delete:
      description: 根据服务器ID删除服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除服务器
      tags:
      - linux-machines
    get:
      description: 根据服务器ID获取服务器详情
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取服务器
      tags:
      - linux-machines
    put:
      consumes:
      - application/json
      description: 根据服务器ID更新服务器信息
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 服务器信息
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/model.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新服务器
      tags:
      - linux-machines
  /api/linux-machines/{id}/connect:
    post:
      description: 测试连接到服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 连接服务器
      tags:
      - linux-machines
  /api/linux-machines/{id}/containers:
    get:
      description: 获取服务器上的Docker容器列表
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.DockerContainer'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取Docker容器列表
      tags:
      - linux-machines
...
swagger: "2.0"